Fix to allow ShaderProgram to use vector<Shader> ok
ShaderProgram kept its shaders as a vector<Shader> object. However, due to how the Shader object's lifecycle works, one of  the Shader's destructor would be called. This caused the OpenGL shader for that Shader obj to be deleted, even tho the OpenGL shader could've been meant to still be used in Nereus. hence move this delete call from the shader destructor to the ShaderProgram's destructor. WARNING: now client code *must* be responsible of calling glDeleteShader once the Shader obj is no longer used...