Fix tessel factor calc to remove gap artefacts
Same edges in different patches had different tessel factors, and so different subdivisions. This resulted in gap artefacts appearing when the mesh was displaced by the heightfield from gerstner waves. This is now fixed as the tessellation factors are now correctly calculated s.t. same edges in different patches have the same tessel factor, and so no gaps appear.